rename `functions` module to `function_like`
