+++ macros/src/function_like.rs
