+++ macros/src/function_like/assert_binop.rs
@@ -3,7 +3,7 @@ use proc_macro::TokenStream;
 use quote::quote;
 use syn::{parse_macro_input, punctuated::Punctuated};
 
-use crate::{construct, functions::log};
+use crate::{construct, function_like::log};
 
 use self::args::Args;
