+++ macros/src/function_like/intern.rs
