+++ macros/src/function_like/log.rs
