+++ macros/src/function_like/log/codegen.rs
