+++ macros/src/function_like/write/args.rs
